requirements.txt
anyio==3.6.2
asyncpg==0.27.0
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
greenlet==2.0.2
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
hvac==1.1.0
idna==3.4
protobuf==4.22.1
psycopg2==2.9.6
pydantic==1.10.7
pyhcl==0.4.4
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
